+++ tests/Integration/MovieDatasetTest.php
@@ -11,6 +11,7 @@
 
 namespace GraphAware\Neo4j\OGM\Tests\Integration;
 
+use GraphAware\Neo4j\OGM\Proxy\LazyCollection;
 use GraphAware\Neo4j\OGM\Tests\Integration\Models\MoviesDemo\Movie;
 use GraphAware\Neo4j\OGM\Tests\Integration\Models\MoviesDemo\Person;
 
@@ -93,5 +94,126 @@ class MovieDatasetTest extends IntegrationTestCase
         $this->em->flush();
         $this->assertGraphExist('(m:Movie {title:"Pirates Of The Caribbean"})<-[:ACTED_IN]-(p:Person {name:"Johnny Depp"})');
         $this->assertCount(1, $this->em->getRepository(Person::class)->findBy(['name' => 'Johnny Depp']));
+    }
+
+    /**
+     * @see https://github.com/graphaware/neo4j-php-ogm/issues/104
+     * @group issue-104
+     */
+    public function testRelationshipReferencesCanBeRemoved()
+    {
+        /** @var Person $person */
+        $person = $this->em->getRepository(Person::class)->findOneBy(['name' => 'Tom Hanks']);
+        /** @var Movie $movie */
+        $movie = $this->em->getRepository(Movie::class)->findOneBy(['title' => 'Cast Away']);
+
+        $castFromPerson = null;
+        foreach ($person->getMovies() as $m) {
+            if ('Cast Away' === $m->getTitle()) {
+                $castFromPerson = $m;
+            }
+        }
+        $this->assertNotNull($castFromPerson);
+        $c = count($person->getMovies());
+        $this->assertEquals(spl_object_hash($castFromPerson), spl_object_hash($movie));
+        $person->getMovies()->removeElement($movie);
+        $movie->getActors()->removeElement($person);
+        $this->assertEquals($c-1, count($person->getMovies()));
+        $this->em->flush();
+
+        $this->assertGraphNotExist('(p:Person {name:"Tom Hanks"})-[:ACTED_IN]->(m:Movie {title:"Cast Away"})');
+    }
+
+    public function testRelationshipReferencesCanBeRemovedTwice()
+    {
+        /** @var Person $person */
+        $person = $this->em->getRepository(Person::class)->findOneBy(['name' => 'Tom Hanks']);
+        /** @var Movie $movie */
+        $movie = $this->em->getRepository(Movie::class)->findOneBy(['title' => 'Cast Away']);
+
+        $castFromPerson = null;
+        foreach ($person->getMovies() as $m) {
+            if ('Cast Away' === $m->getTitle()) {
+                $castFromPerson = $m;
+            }
+        }
+        $this->assertNotNull($castFromPerson);
+        $c = count($person->getMovies());
+        $this->assertEquals(spl_object_hash($castFromPerson), spl_object_hash($movie));
+        $person->getMovies()->removeElement($movie);
+        $movie->getActors()->removeElement($person);
+        $this->assertEquals($c-1, count($person->getMovies()));
+        $this->em->flush();
+        $this->em->flush();
+
+        $this->assertGraphNotExist('(p:Person {name:"Tom Hanks"})-[:ACTED_IN]->(m:Movie {title:"Cast Away"})');
+    }
+
+    public function testRelationshipReferenceCanBeReAddedAndRemoved()
+    {
+        /** @var Person $person */
+        $person = $this->em->getRepository(Person::class)->findOneBy(['name' => 'Tom Hanks']);
+        /** @var Movie $movie */
+        $movie = $this->em->getRepository(Movie::class)->findOneBy(['title' => 'Cast Away']);
+
+        $castFromPerson = null;
+        foreach ($person->getMovies() as $m) {
+            if ('Cast Away' === $m->getTitle()) {
+                $castFromPerson = $m;
+            }
+        }
+        $this->assertNotNull($castFromPerson);
+        $c = count($person->getMovies());
+        $this->assertEquals(spl_object_hash($castFromPerson), spl_object_hash($movie));
+        $person->getMovies()->removeElement($movie);
+        $movie->getActors()->removeElement($person);
+        $this->assertEquals($c-1, count($person->getMovies()));
+        $this->em->flush();
+        $this->assertGraphNotExist('(p:Person {name:"Tom Hanks"})-[:ACTED_IN]->(m:Movie {title:"Cast Away"})');
+        $person->getMovies()->add($movie);
+        $movie->getActors()->add($person);
+        $this->em->flush();
+        $this->assertGraphExist('(p:Person {name:"Tom Hanks"})-[:ACTED_IN]->(m:Movie {title:"Cast Away"})');
+        $person->getMovies()->removeElement($movie);
+        $movie->getActors()->removeElement($person);
+        $this->em->flush();
+        $this->assertGraphNotExist('(p:Person {name:"Tom Hanks"})-[:ACTED_IN]->(m:Movie {title:"Cast Away"})');
+    }
+
+    public function testRelationshipReferenceCanBeRemovedAfterNewCreation()
+    {
+        /** @var Person $person */
+        $person = $this->em->getRepository(Person::class)->findOneBy(['name' => 'Tom Hanks']);
+        $movie = new Movie('Super Movie');
+        $person->getMovies()->add($movie);
+        $movie->getActors()->add($person);
+        $person->getMovies()->first();
+        $this->em->flush();
+        $this->assertGraphExist('(p:Person {name:"Tom Hanks"})-[:ACTED_IN]->(m:Movie {title:"Super Movie"})');
+        $person->getMovies()->removeElement($movie);
+        $movie->getActors()->removeElement($person);
+        $this->em->flush();
+        $this->assertGraphNotExist('(p:Person {name:"Tom Hanks"})-[:ACTED_IN]->(m:Movie {title:"Super Movie"})');
+    }
+
+    public function testThatGetterShouldNotBeCalledToTriggerLazyLoading()
+    {
+        /** @var Person $person */
+        $person = $this->em->getRepository(Person::class)->findOneBy(['name' => 'Tom Hanks']);
+        $this->assertInstanceOf(LazyCollection::class, $person->movies);
+
+        /** @var Movie $movie */
+        $movie = $this->em->getRepository(Movie::class)->findOneBy(['title' => 'The Matrix']);
+        $this->assertInstanceOf(LazyCollection::class, $movie->actors);
+    }
+
+    public function testDegreeOfNodeIsReturned()
+    {
+        $degree = $this->client->run('MATCH (n:Person {name:"Tom Hanks"}) RETURN size((n)-[:ACTED_IN]->()) AS c')->firstRecord()->get('c');
+        /** @var Person $person */
+        $person = $this->em->getRepository(Person::class)->findOneBy(['name' => 'Tom Hanks']);
+        $this->assertEquals($degree, $person->getMovies()->count());
+
+
     }
 }